+++ common/src/nativeMain/kotlin/com/powersync/db/NativeConnectionFactory.kt
@@ -1,16 +1,24 @@
 package com.powersync.db
 
 import androidx.sqlite.SQLiteConnection
+import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PersistentConnectionFactory
 import com.powersync.PowerSyncException
+import com.powersync.resolvePowerSyncLoadableExtensionPath
 import com.powersync.sqlite.Database
 
 /**
  * A [PersistentConnectionFactory] implementation delegating to static `sqlite3_` invocations through cinterop.
  */
-public abstract class NativeConnectionFactory: PersistentConnectionFactory {
-    override fun openConnection(path: String, openFlags: Int): SQLiteConnection {
-        val extensionPath = powersyncLoadableExtensionPath()
+public abstract class NativeConnectionFactory : PersistentConnectionFactory {
+    @OptIn(ExperimentalPowerSyncAPI::class)
+    override fun openConnection(
+        path: String,
+        openFlags: Int,
+    ): SQLiteConnection {
+        // On some platforms, most notably watchOS, there's no dynamic extension loading and the core extension is
+        // registered via sqlite3_auto_extension.
+        val extensionPath = resolvePowerSyncLoadableExtensionPath()
         val db = Database.open(path, openFlags)
 
         if (extensionPath != null) {
@@ -25,14 +33,5 @@ public abstract class NativeConnectionFactory: PersistentConnectionFactory {
         return db
     }
 
-    override fun openInMemoryConnection(): SQLiteConnection {
-        return openConnection(":memory:", 0x02)
-    }
-
-    /**
-     * If the core extension should be loaded as a dynamic library, returns its path.
-     *
-     * Otherwise, installs the core extension as a static extension and returns null.
-     */
-    protected abstract fun powersyncLoadableExtensionPath(): String?
+    override fun openInMemoryConnection(): SQLiteConnection = openConnection(":memory:", 0x02)
 }
